camera: remove unneeded ois params in Camera2
'ois-mode-values' and 'ois-supported' keys are only used to report back to
framework. Besides, 'ois-supported' is already set to true.

ois mode can be set with the 'ois' key which can take one of those values:
still, still_zoom, video, sine_x, sine_y or center.

However, it is already handled automatically. When 'ois' key is set to
'still' (the default), ois mode is set to 'still' (100) when
RecordingHint=false and set to 'video' (102) when RecordingHint=true.

You can check this using logcat:

[CAM_ID(0)][ParametersBack]-DEBUG(setParameters):newRecordingHint : false, newEffectRecordingHint : 0
[CAM_ID(0)][ParametersBack]-(setOISMode[3976])set OIS Mode = 100

or

[CAM_ID(0)][ParametersBackDummy]-DEBUG(setParameters):newRecordingHint : true, newEffectRecordingHint : 0
[CAM_ID(0)][ParametersBack]-(setOISMode[3976])set OIS Mode = 102

Change-Id: I1f84e3e1903e8405e0d4b87b6acf0c19aa77ea5d