hero-common: overlay: Improve Snap image quality
When using Camera API 2, quality is not really good and 3DNR is not activated
adding a lot of noise in low-light conditions pictures so use Camera API 1.

Dual pixel autofocus is not activated by default so enable it
(phase-af=on).

Realtime hdr and drc show great results (rt-hdr=auto and drc=true, thanks to
Jesse Chan for the track).

Picture quality seems to be similar to stock rom using samsung camera app
(sometimes better for my taste).

However, camera locks up quite often when using Camera API 1 so it cannot be
a sustainable solution.

Change-Id: Ibc56021b5b7c4fe5abd573c6cdd309daaf9f69f3